requirements.txt
streamlit
numpy
matplotlib
torch
torchvision
Pillow
requests
pandas
nltk
PyPDF2
pillow